Revert "new change like remove username and replace with email"
