+++ src/Admin/CompanyDelete.jsx
@@ -6,7 +6,7 @@ import axios from "axios";
 import { Button, Container, Hidden } from "@mui/material";
 import AddIcon from "@mui/icons-material/Add";
 import { Fab, Paper, styled } from "@mui/material";
-import country from "../Api/countryCityState.json"
+import country from "../Api/countriess.json"
 import DeleteSweepOutlinedIcon from '@mui/icons-material/DeleteSweepOutlined';
 import env from "react-dotenv";
