+++ src/Admin/CompanyEdit.jsx
@@ -6,7 +6,7 @@ import axios from "axios";
 import { Button, Container, Tooltip, styled } from "@mui/material";
 import AddIcon from "@mui/icons-material/Add";
 import { Fab } from "@mui/material";
-import country from "../Api/countryCityState.json";
+import country from "../Api/countriess.json";
 import { ToastContainer, toast } from "react-toastify";
 import EditNoteOutlinedIcon from "@mui/icons-material/EditNoteOutlined";
