+++ src/Employee/EmployeeCreate.jsx
@@ -6,7 +6,7 @@ import Modal from "@mui/material/Modal";
 import { ToastContainer, toast } from "react-toastify";
 import { Button, Container } from "@mui/material";
 import env from "react-dotenv";
-import country from "../Api/countryCityState.json";
+import country from "../Api/countriess.json";
 import employeeRole from "../jsonlist/employeeRole.json"
 import {
   validatePhoneNumber,
