+++ src/checkwriter/Screen1.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Screen1 = () => {
+  return (
+    <div>
+      screen one
+    </div>
+  )
+}
+
+export default Screen1
