+++ src/company/ProjectEdit.jsx
@@ -3,7 +3,7 @@ import Box from "@mui/material/Box";
 import Modal from "@mui/material/Modal";
 import { useState, useEffect, useContext } from "react";
 import axios from "axios";
-import country from "../Api/countryCityState.json";
+import country from "../Api/countriess.json";
 // import states from "../Api/states.json"
 import { ToastContainer, toast } from "react-toastify";
 import "react-toastify/dist/ReactToastify.css";
